1) Συνδέθηκαν τα πραγματικά δεδομένα από την firebase στον RV και στην detail activities 2) Offline λειτουργία firebase
