+++ app/src/main/java/greek/dev/challenge/charities/views/CharitiesResultsActivity.java
@@ -1,7 +1,10 @@
 package greek.dev.challenge.charities.views;
 
+import android.arch.lifecycle.Observer;
+import android.arch.lifecycle.ViewModelProviders;
 import android.content.Context;
 import android.content.Intent;
+import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.LinearLayoutManager;
@@ -11,7 +14,16 @@ import android.view.View;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
+import com.google.firebase.database.ChildEventListener;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.storage.FirebaseStorage;
+import com.google.firebase.storage.StorageReference;
+
 import java.util.ArrayList;
+import java.util.List;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -32,6 +44,14 @@ public class CharitiesResultsActivity extends AppCompatActivity implements Resul
 
     private ResultsAdapter mCharitiesAdapter;
 
+    private FirebaseDatabase mFirebaseDatabase;
+    private DatabaseReference mCharitiesDatabaseReference;
+    private ChildEventListener mChildEventListener;
+    private FirebaseStorage mFirebaseStorage;
+    private StorageReference mCharitiesPhotosStorageReference;
+    private MainViewModel viewModel;
+
+    static boolean calledFirebaseAlready = false;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -50,7 +70,34 @@ public class CharitiesResultsActivity extends AppCompatActivity implements Resul
 
         pb_loading_indicator = (ProgressBar) findViewById(R.id.pb_loading_indicator);
 
-        // TODO THE CODE BELLOW IS FOR TESTING. has to be removed to implement it using the real data.
+
+        //get reference to specific part of database - messages with mMessagesDatabaseReference
+
+        if (!calledFirebaseAlready)
+        {
+            FirebaseDatabase.getInstance().setPersistenceEnabled(true);
+            calledFirebaseAlready = true;
+        }
+        mFirebaseDatabase = FirebaseDatabase.getInstance();
+        mFirebaseStorage = FirebaseStorage.getInstance();
+
+        mCharitiesDatabaseReference = mFirebaseDatabase.getReference().child("charities");
+
+        // mCharitiesPhotosStorageReference = mFirebaseStorage.getReference().child("charities_photos");
+
+        viewModel = ViewModelProviders.of(this).get(MainViewModel.class);
+
+        viewModel.getCharitiesList().observe(CharitiesResultsActivity.this, new Observer<ArrayList<Charity>>() {
+            @Override
+            public void onChanged(@Nullable ArrayList<Charity> charitiesList) {
+                mCharitiesAdapter.setSeriesResults(charitiesList);
+                Log.v("main", charitiesList.get(charitiesList.size()-1).toString());
+                Log.v("main", String.valueOf(charitiesList.size()));
+            }
+        });
+
+
+     /*   // TODO THE CODE BELLOW IS FOR TESTING. has to be removed to implement it using the real data.
         // --------------------------------------------------------
         Charity test1 = new Charity(1, "Charity Name 1", "lalala1", "5", "test", "test", "10", "10", "lala", "test");
         Charity test2 = new Charity(2, "Charity Name 2", "lalala2", "6", "test2", "test2", "102", "102", "lal2a", "test2");
@@ -59,10 +106,68 @@ public class CharitiesResultsActivity extends AppCompatActivity implements Resul
         testing.add(test1);
         testing.add(test2);
         // ----------------------------------------------------------
+*/
 
-        mCharitiesAdapter.setSeriesResults(testing);
     }
+    @Override
+    protected void onResume() {
+        super.onResume();
+        // an exei ksanakatevasei ta dedomena den xreiazetai na ta ksanatraviksoume
+        if (!viewModel.fetched){
+            attachDatabaseReadListener();
+        }
+    }
+    private void attachDatabaseReadListener() {
+        if (mChildEventListener == null) {
+            mChildEventListener = new ChildEventListener() {
+                @Override
+                public void onChildAdded(DataSnapshot dataSnapshot, String s) {
+                    //here we get the results from the firebase db
+                    Charity charity = dataSnapshot.getValue(Charity.class);
+                    viewModel.addCharity(charity);
+                    viewModel.fetched = true;
+                }
 
+                @Override
+                public void onChildChanged(DataSnapshot dataSnapshot, String s) {
+
+                }
+
+                @Override
+                public void onChildRemoved(DataSnapshot dataSnapshot) {
+
+                }
+
+                @Override
+                public void onChildMoved(DataSnapshot dataSnapshot, String s) {
+
+                }
+
+                @Override
+                public void onCancelled(DatabaseError databaseError) {
+
+                }
+            };
+            mCharitiesDatabaseReference.addChildEventListener(mChildEventListener);
+
+        }
+    }
+    private void detachDatabaseReadListener() {
+        if (mChildEventListener != null) {
+            mCharitiesDatabaseReference.removeEventListener(mChildEventListener);
+            mChildEventListener = null;
+        }
+    }
+    @Override
+    protected void onPause() {
+        super.onPause();
+        onSignedOutCleanup();
+    }
+    //not signed out now, but a cleanup is required onPause, so not to get duplicate EventListeners
+    private void onSignedOutCleanup(){
+        detachDatabaseReadListener();
+
+    }
 
     private void showCharitiesList() {
         tv_error_message.setVisibility(View.INVISIBLE);
@@ -80,7 +185,8 @@ public class CharitiesResultsActivity extends AppCompatActivity implements Resul
     public void onClick(Charity selectedCharity) {
         Context context = this;
         Intent intent = new Intent(context, CharityDetails.class);
-        intent.putExtra(Intent.EXTRA_TEXT, ""+selectedCharity.getId());
+        intent.putExtra("charity",selectedCharity);
+        //intent.putExtra(Intent.EXTRA_TEXT, ""+selectedCharity.getId());
         startActivity(intent);
     }
 }
