--- app/src/main/java/greek/dev/challenge/charities/views/GridItemActivity1.java
@@ -1,20 +1,0 @@
-package greek.dev.challenge.charities.views;
-
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
-import greek.dev.challenge.charities.R;
-
-/**
- * Created by eirinimitsopoulou on 18/12/2017.
- */
-
-public class GridItemActivity1 extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.results_recycler_view);
-    }
-}
