+++ app/src/main/java/greek/dev/challenge/charities/views/MainViewModel.java
@@ -12,16 +12,17 @@ import java.util.List;
 import greek.dev.challenge.charities.model.Charity;
 
 public class MainViewModel extends ViewModel {
-    public MutableLiveData<List<Charity>> charitiesList;
-
-    public LiveData<List<Charity>> getCharitiesList() {
+    public MutableLiveData<ArrayList<Charity>> charitiesList;
+    //an exei eidi katebasei ta dedomena gia na min xreiazetai na ksanakatevoun
+    public boolean fetched = false;
+    public LiveData<ArrayList<Charity>> getCharitiesList() {
         if (charitiesList == null){
-            charitiesList = new MutableLiveData<List<Charity>>();
+            charitiesList = new MutableLiveData<ArrayList<Charity>>();
         }
         return charitiesList;
     }
     public void addCharity(Charity charity){
-         List<Charity> charityListTemp;
+         ArrayList<Charity> charityListTemp;
          charityListTemp = charitiesList.getValue();
             if (charityListTemp == null) {
                 charityListTemp = new ArrayList<Charity>();
